default enable const-vbv
it was default in tune grain only, and the commit that introduced it says it was added for tune grain, but it also works in general settings.
the quality is the same *level*, I don't simply just say "the same" because "plain vbv" is non-deterministic when it's engaged, so there's no way to be sure.
and if you are very worried about quality you probably shouldn't use vbv to begin with. heck, maybe don't even use a lossy codec.